[Refactor] (#11)
* Initial commit

* Initial commit

* [feat] swagger config 추가

* [Feat] 공통 응답 구현(예외 메시지 처리 핸들러는 미구현)

* 공통 시간 Entity 객체

* [docs] properties -> yml, db 환경 변수 설정

* [Build] db, swagger 의존성 추가

* [Feat] 더미 데이터 및 DB에 initialize 구현

* [Feat] Food 타입과 엔터티 구현

* [Feat] FoodFeedback 엔터티 구현

* [Feat] Menu 엔터티 구현

* [Feat] 인기 상위 5, 하위 5개 메뉴 조회 API 구현

* @EntityListeners(AuditingEntityListener.class 위한 JpaConfig 구현

* 코드 컨벤션

* FoodFeedBack -> FoodMenu로 변경

* cors config 구현

* [Feat] 랭킹 변화 -5 ~ 5 사이 반환하도록 임의 설정

* .

* kafka 추가

* 데이터 더미를 위한 코드 삭제

* [Refactor] ERD에 맞게 수정

* Food 필드 FoodName -> Name으로 변경

* feat: update create demodata

* feat update create demo dataset

* feat: update base url point

* feat: update test get method & db config

* .

* .

* application.yml 추척 무시하도록 변경

* 필요없는 클래스 제거 (데이터 초기화 하는 클래스)

* Dataset controller 코드 컨벤션 및 swagger 설명 추가

* MSA 팀별 URL 상수처리

* [Feat] 주간, 월간 통계 조회 API 구현

* [Feat] 주간, 월간 통계 조회 API 구현

* [Refactor] rank가 statistics안에 들어가도록 변경

* [Feat] 주간,월간 통계 조회 API 구현

* [Refactor]RankMenuResponseDto, RankMenuResponse 클래스명에서 CombinedRankMenuResponseDto를 통해 명확하도록 변경

* [Feat] Rank entity 추가

* [Feat] 스케줄링 작업을 통해 피드백을 통해 음식 rank 계산하는 기능 구현

* [Feat] Rank entity 추가

* PeriodType 네임 변경

* PeriodType 네임 변경

* 패키지 이동

* 클래스 제거

* .

* 팀원의 변경사항 merge

* csv로 데이터 초기화하던 코드 제거

* 팀원의 변경사항 merge

* [Refactor] 응답값에 들어가는 rank 값은 실제 rank가 아닌 순위를 보여주도록 변경

* FeedBack의 score를 double -> Integer로 변경

* 직접 데이터를 넣기 위해서 BaseTimeEntity의 Listener를 수정

* 주간 피드백을 통한 rank 초기화 api 구현

* 직접 데이터를 넣기 위해서 BaseTimeEntity의 Listener를 수정

* RankImpl의 책임을 분리

* 이전 순위가 없으면 0을 반환.

* FoodRank를 기록날, 기간타입을 통한 조회 메서드 추가

* FeedBack의 score를 double로 변경

* scheduler가 아닌 service를 호출하도록 변경

---------

+++ src/main/java/com/example/SchoolLunchReport/dummy/dataset/dto/FeedBackDTO.java
@@ -1,10 +1,12 @@
 package com.example.SchoolLunchReport.dummy.dataset.dto;
+
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class FeedBackDTO {
+
     private Long foodMenuId;
-    private Integer score;
+    private Double score;
 }
